+++ changelog.md
@@ -8,6 +8,7 @@ This file contains all the notable changes done to the Ballerina GraphQL package
 - [[#1000] Enum Support](https://github.com/ballerina-platform/ballerina-standard-library/issues/1000)
 - [[#999] Map Support](https://github.com/ballerina-platform/ballerina-standard-library/issues/999)
 - [[#1091] Compiler plugin validations for GraphQL services](https://github.com/ballerina-platform/ballerina-standard-library/issues/1091)
+- [[#1001] Support Union Types](https://github.com/ballerina-platform/ballerina-standard-library/issues/1001)
 
 ## [0.2.0-alpha6] - 2021-04-02