+++ changelog.md
@@ -10,6 +10,9 @@ This file contains all the notable changes done to the Ballerina GraphQL package
 - [[#1091] Compiler plugin validations for GraphQL services](https://github.com/ballerina-platform/ballerina-standard-library/issues/1091)
 - [[#1001] Support Union Types](https://github.com/ballerina-platform/ballerina-standard-library/issues/1001)
 
+### Fixed
+- [[#1250] Fix Hanging the Service when Returning Array](https://github.com/ballerina-platform/ballerina-standard-library/issues/1250)
+
 ## [0.2.0-alpha6] - 2021-04-02
 
 ### Added